feat: Introduce score to quickly assess quality of an assignment run.
The score is the sum of fulfilled priorities across all participants.

Example: Participants had to specify six priorities and should be assigned to two activities. If somebody gets their first and fifth priority fulfilled they are scored with 6. If somebody only gets their second priority fulfilled they get a score of 8, because missing assignments are scored with the number of priorities.